Refactor Github actions and workflows
Removed a number of Github workflows and actions that are no longer necessary. In place of these, two new workflows have been added: 'release' and 'publish'. The 'release' workflow generates a changelog and releases a new tag, and the 'publish' workflow is triggered on release creation to publish the package to NPM.